Added required library.
If you want to parse an org document as an AST then the library
"Org-element" is required. The "Org-element" library is shipped
with the standard version of emacs but it does need to be imported
manually if you want it to be accessible.I have added syntax for that
library to be imported.